Fix IndexError in prepare_heatmap_data when filter values missing
- Fix beta_filter type from int to float (4 → 4.0) in OMLParameterGrid
- Refactor prepare_heatmap_data to handle missing filter combinations gracefully
- Add validation in plot_gennorm_suite to skip plots when filter values not in data

The IndexError occurred when the list comprehension returned an empty list
and code tried to access [0]. This happens when filter values (beta_filter
or support_filter) don't exist in the results data.

+++ oml_utils.py
@@ -72,7 +72,7 @@ class OMLParameterGrid:
     treatment_effects: List[float] = field(default_factory=lambda: [0.01, 0.1, 0.5, 1.0, 3.0, 10])
     treatment_coefs: List[float] = field(default_factory=lambda: [-0.002, 0.05, -0.43, 1.56])
     outcome_coefs: List[float] = field(default_factory=lambda: [0.003, -0.02, 0.63, -1.45])
-    beta_filter: int = 4
+    beta_filter: float = 4.0
     support_filter: int = 10
 
     @classmethod
